backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
boto3==1.35.43
botocore==1.35.43
certifi==2025.1.31
cffi==1.17.1
decouple==0.0.7
Django==5.1.1
django-appconf==1.0.6
django-compressor==4.5.1
django-environ==0.11.2
django-grappelli==4.0.1
django-minio-storage==0.5.7
django-redis==5.4.0
django-storages==1.14.4
gunicorn==23.0.0
jmespath==1.0.1
minio==7.2.15
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-decouple==3.8
rcssmin==1.1.2
redis==5.2.0
rjsmin==1.2.2
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
